extracted FilePermissions into separate component
