+++ client/src/pages/task/TaskConfigurationPage.tsx
@@ -1,6 +1,5 @@
 import {
   EditOutlined,
-  InfoCircleFilled,
   PoweroffOutlined,
   SyncOutlined
 } from '@ant-design/icons'
@@ -9,23 +8,17 @@ import {
   Button,
   Card,
   Checkbox,
-  Col,
   Collapse,
   Empty,
-  List,
-  Row,
   Space,
-  Switch,
-  Tooltip
+  Switch
 } from 'antd'
 import { CheckboxChangeEvent } from 'antd/lib/checkbox'
-import { JSONSchema6 } from 'json-schema'
 import React, { useState } from 'react'
 import ReactMarkdown from 'react-markdown'
 import { Link } from 'react-router-dom'
 import AsyncPlaceholder from '../../components/AsyncPlaceholder'
 import EditableMarkdown from '../../components/EditableMarkdown'
-import JsonSchemaEditButton from '../../components/JsonSchemaEditButton'
 import StartSubmissionEvaluationButton from '../../components/StartSubmissionEvaluationButton'
 import useIdParam from '../../hooks/useIdParam'
 import {
@@ -41,40 +34,13 @@ import EditEvaluationPage from '../evaluation/EditEvaluationPage'
 import IdeSettingsForm, {
   IdeSettingsModel
 } from '../../components/IdeSettingsForm'
+import FilePermissions from "../../components/FilePermissions";
 
 const { Panel } = Collapse
 
-const renderFilePattern = (pattern: string) => (
-  <List.Item>
-    <code>{pattern}</code>
-  </List.Item>
-)
 
-const filePatternSchema: JSONSchema6 = {
-  type: 'array',
-  items: { type: 'string' }
-}
 
-const filePatternHelp = (
-  <Alert
-    style={{ marginBottom: 16 }}
-    message={
-      <>
-        File patterns use the Ant pattern syntax. For more information refer to
-        the{' '}
-        <a
-          href="https://ant.apache.org/manual/dirtasks.html#patterns"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          official documentation
-        </a>
-        .
-      </>
-    }
-    type="info"
-  />
-)
+
 
 const TaskConfigurationPage: React.FC<{ editable: boolean }> = ({
   editable
@@ -164,7 +130,7 @@ const TaskConfigurationPage: React.FC<{ editable: boolean }> = ({
       </Collapse>
       <Collapse>
         <Panel
-          header="Files"
+          header="File Permissions"
           extra={
             <>
               <Link
@@ -223,80 +189,7 @@ const TaskConfigurationPage: React.FC<{ editable: boolean }> = ({
               showIcon
             />
           ) : null}
-          <Row gutter={16}>
-            <Col span={12}>
-              <List
-                size="small"
-                header={
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'space-between',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <span>
-                      <b>Hidden files</b>{' '}
-                      <Tooltip
-                        title={
-                          'Patterns of files that should be hidden from students. Matching files are only included for evaluation. If matching files are created by students, they are ignored for evaluation.'
-                        }
-                        placement="bottom"
-                      >
-                        <InfoCircleFilled />
-                      </Tooltip>
-                    </span>
-                    <JsonSchemaEditButton
-                      title="Edit hidden files"
-                      extraContent={filePatternHelp}
-                      schema={filePatternSchema}
-                      value={task.hiddenFiles}
-                      onSubmit={updater('hiddenFiles')}
-                    />
-                  </div>
-                }
-                bordered
-                dataSource={task.hiddenFiles}
-                renderItem={renderFilePattern}
-              />
-            </Col>
-            <Col span={12}>
-              <List
-                size="small"
-                header={
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'space-between',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <span>
-                      <b>Protected files</b>{' '}
-                      <Tooltip
-                        title={
-                          'Patterns of files that should be read-only. Students will be able to see matching files but modifications are ignored for evaluation. Non-existent files can be protected to prevent their creation.'
-                        }
-                        placement="bottom"
-                      >
-                        <InfoCircleFilled />
-                      </Tooltip>
-                    </span>
-                    <JsonSchemaEditButton
-                      title="Edit protected files"
-                      extraContent={filePatternHelp}
-                      schema={filePatternSchema}
-                      value={task.protectedFiles}
-                      onSubmit={updater('protectedFiles')}
-                    />
-                  </div>
-                }
-                bordered
-                dataSource={task.protectedFiles}
-                renderItem={renderFilePattern}
-              />
-            </Col>
-          </Row>
+          <FilePermissions task={task} updater={updater} />
         </Panel>
       </Collapse>
       <Collapse>
